# trie, tests/fuzzers: implement a stacktrie fuzzer + stacktrie fixes