This PR contains an almost functional stacktrie-fuzzer, which fuzzes stacktrie against the regular trie. The problem with the fuzzer is that it immediately hits crashes. 

Example of one such vector has been added in stacktrie_test.go, and it would be good if we got that resolved. 

Also, while implementing the fuzzer, I found that theoretical tries with roots small enough to be `<32 bytes` do not get committed to the db. This has been fixed, and a testcase for it has been added. 

@gballet please help me figure out why the test `TestUpdate` fails, and also verify the fix for `Commit` (cc @rjl493456442 )