@holiman 2020-11-08
I pushed a fix which makes the fuzzer work (by always using 32-byte keys). The PR stil contains a failing testcase though, which at least in _theory_ points at an error in stacktrie. So we should either fix it, or properly ensure that it can never happen in practice, and if so, document the behaviour.