@holiman 2020-11-09
Thanks to @rjl493456442 for solving the issue with stacktrie! The fix is now included in this PR, and does not seem to meaningfully impact the performance: 
```
name                       old time/op    new time/op    delta
DeriveSha200/std_trie-6       536µs ± 1%     565µs ±19%   ~     (p=0.841 n=5+5)
DeriveSha200/stack_trie-6     447µs ± 1%     448µs ± 3%   ~     (p=1.000 n=4+5)

name                       old alloc/op   new alloc/op   delta
DeriveSha200/std_trie-6       274kB ± 0%     274kB ± 0%   ~     (p=0.095 n=5+5)
DeriveSha200/stack_trie-6    52.5kB ± 0%    52.5kB ± 0%   ~     (p=0.135 n=5+5)

name                       old allocs/op  new allocs/op  delta
DeriveSha200/std_trie-6       2.67k ± 0%     2.67k ± 0%   ~     (all equal)
DeriveSha200/stack_trie-6     1.07k ± 0%     1.07k ± 0%   ~     (all equal)

```